Attempt API image failed
I tried to get images from the API to show on my site, but i did not succeed. I have searched around on the internet, and watched the videos from the school, but still don't get it. I looked up in a book i have, which i believe lead me closer to make it work. Still i need some more research to do.