Update 4
It took some time, but i have finally managed to fetch some images from the API. I have used a test file before i changed the code in the assignment. I tried to connect to the v2 API, but did not succeed. It worked with V1, so i used that one.

On the main page, i have added all 10 products and sorted them in rows with 5 games per row. I have also made them link to the product/index.html page. Now i will work with getting the right information there based on which game the user clicked on.