Mini update
I have started working on sending information over pages. The plan is to figure out which game the user is clicking on to send the user information about the game in the product page. SO far i am just cheking waht iformation i can get from each product. So far so good