Update 5
I have successfully transfered information between the pages. I can now click on a product in the index page, and it will transfer me to the product page for that specific product. The information thers is correct according to the API.

Last step is to work on the cart and checkout page. I will do that tomorrow